@@ -8,7 +8,7 @@ authors = ["Sebastian Thiel <sebastian.thiel@icloud.com>"]
 edition = "2021"
 license = "MIT OR Apache-2.0"
 version = "0.45.0"
-rust-version = "1.74"
+rust-version = "1.82"
 default-run = "gix"
 include = ["src/**/*", "/build.rs", "LICENSE-*", "README.md"]
 resolver = "2"
@@ -174,7 +174,6 @@ terminal_size = "0.4.2"
 # Avoid pre-compiled binaries, see https://github.com/serde-rs/serde/issues/2538 and https://github.com/serde-rs/serde/pull/2590
 serde_derive = ">=1.0.185"
 
-once_cell = "1.21.3"
 document-features = { version = "0.2.0", optional = true }
 [profile.dev.package]
 insta.opt-level = 3
