@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 description = "archive generation from of a worktree stream"
 authors = ["Sebastian Thiel <sebastian.thiel@icloud.com>"]
 edition = "2021"
-rust-version = "1.70"
+rust-version = "1.82"
 include = ["src/**/*", "LICENSE-*"]
 
 [lib]
