@@ -2,7 +2,7 @@ use gix_fsck::Connectivity;
 use gix_hash::ObjectId;
 use gix_hashtable::HashMap;
 use gix_object::Kind;
-use gix_testtools::once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
 use crate::hex_to_id;
 
