@@ -5,14 +5,14 @@ use std::{
 };
 
 use bstr::{BStr, BString, ByteSlice};
-use once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
 /// Other places to find Git in.
 #[cfg(windows)]
-pub(super) static ALTERNATIVE_LOCATIONS: Lazy<Vec<PathBuf>> =
-    Lazy::new(|| locations_under_program_files(|key| std::env::var_os(key)));
+pub(super) static ALTERNATIVE_LOCATIONS: LazyLock<Vec<PathBuf>> =
+    LazyLock::new(|| locations_under_program_files(|key| std::env::var_os(key)));
 #[cfg(not(windows))]
-pub(super) static ALTERNATIVE_LOCATIONS: Lazy<Vec<PathBuf>> = Lazy::new(Vec::new);
+pub(super) static ALTERNATIVE_LOCATIONS: LazyLock<Vec<PathBuf>> = LazyLock::new(Vec::new);
 
 #[cfg(windows)]
 fn locations_under_program_files<F>(var_os_func: F) -> Vec<PathBuf>
@@ -101,7 +101,7 @@ pub(super) const EXE_NAME: &str = "git";
 /// Invoke the git executable to obtain the origin configuration, which is cached and returned.
 ///
 /// The git executable is the one found in `PATH` or an alternative location.
-pub(super) static GIT_HIGHEST_SCOPE_CONFIG_PATH: Lazy<Option<BString>> = Lazy::new(exe_info);
+pub(super) static GIT_HIGHEST_SCOPE_CONFIG_PATH: LazyLock<Option<BString>> = LazyLock::new(exe_info);
 
 // There are a number of ways to refer to the null device on Windows, but they are not all equally
 // well supported. Git for Windows rejects `\\.\NUL` and `\\.\nul`. On Windows 11 ARM64 (and maybe
@@ -191,7 +191,7 @@ fn first_file_from_config_with_origin(source: &BStr) -> Option<&BStr> {
 /// errors during execution.
 pub(super) fn install_config_path() -> Option<&'static BStr> {
     let _span = gix_trace::detail!("gix_path::git::install_config_path()");
-    static PATH: Lazy<Option<BString>> = Lazy::new(|| {
+    static PATH: LazyLock<Option<BString>> = LazyLock::new(|| {
         // Shortcut: Specifically in Git for Windows 'Git Bash' shells, this variable is set. It
         // may let us deduce the installation directory, so we can save the `git` invocation.
         #[cfg(windows)]
