@@ -4,7 +4,7 @@ use std::{collections::HashMap, path::PathBuf, str::FromStr};
 use gix_object::{bstr, bstr::BStr};
 use gix_ref::bstr::{BString, ByteSlice};
 use gix_revision::spec::Kind;
-use gix_testtools::once_cell::sync::Lazy;
+use std::sync::LazyLock;
 
 const FIXTURE_NAME: &str = "make_rev_spec_parse_repos.sh";
 static BASELINE: Lazy<HashMap<PathBuf, HashMap<BString, Option<gix_revision::Spec>>>> = Lazy::new(|| {
